Fix "goto skips variable initialization" errors
This "fixes" the problem by moving certain variable declarations to the
top top of the file, and by replacing a function pointer with a bunch
of ifs and separate function calls.

While this is ugly, it is no uglier than the rest of those files are
already.  Curiously, it returns the code to something which more
resembles the original, mechanically translated FORTRAN code.  The
code that triggered the errors was in fact changed by hand by me
because I couldn't stand the original. :)